Added Membership completion step, tested
